Treats LeftJoin predicates in SelectMerge
LeftJoin is a special case of a predicate since it requires
additional relations to be presence in order to be applied.
This is represented by the Extended Eligible List that must
be updated appropriately when merging two SelectBlocks.

Moreover, in SPARQL, LeftJoin predicate also has an additional
filter expression that must also be treated while creating the
QueryBlocks.

The treatment is not complete. Some FIXME added for that purpose.